project/requirements.txt
Django>=4.2,<5
djangorestframework>=3.14
djangorestframework-simplejwt>=5.2
mysqlclient>=2.1
celery[redis]>=5.3
redis>=4.5
python-dotenv>=1.0
requests>=2.31
beautifulsoup4>=4.12
drf-yasg>=1.21
django-cors-headers>=3.14
django-ratelimit>=3.0
django-filter
gunicorn>=20.1
pytest>=7.4
pytest-django>=4.5
playwright>=1.40
black
flake8
